Aramco_AIS_Repo_updated_with_src/Aramco_AIS_Repo/P4_Energy_Setpoint_Optimization/requirements.txt
pandas>=1.5
numpy>=1.21
scikit-learn>=1.2
ax-platform>=0.3
botorch>=0.9
torch>=2.0
matplotlib>=3.7
mlflow>=2.3
pyyaml>=5.4